Add favorite check, toggle label & styles
